Implemented RPC and IoT Layer

+++ dat110-project1/src/no/hvl/dat110/system/sensor/SensorDevice.java
@@ -8,18 +8,16 @@ public class SensorDevice {
 	public static void main(String[] args) {
 		
 		System.out.println("Sensor server starting ...");
-		
 		SensorImpl sensor = new SensorImpl();
-		
+
 		RPCServer sensorserver = new RPCServer(Common.SENSORPORT);
-		
-	    sensorserver.register(1,sensor);
-		
+
+		sensorserver.register((byte)1,sensor);
+
 		sensorserver.run();
-		
+
 		sensorserver.stop();
-		
+
 		System.out.println("Sensor server stopping ...");
-		
 	}
 }
